==> 01_prepare_preds_vars/4_get_simulated_vegetation_stats.Rmd <==
---
title: "R Notebook - Analyse iLand output"
output: html_notebook
---

After having run the iLand simulations for the 12 clusters the resulting .sqlite files need to be evaluated. Goal is to see how the tree development of each cluster develops.

# 1. Load data from .sqlite files

To be able to load the data, fist a connection to the .sqlite file needs to be established. For this the library "RSQLite" is needed.

```{r librarys}
#install.packages("RSQLite")
library(RSQLite)
library(dplyr)
library(tidyr)
library(ggpubr)
library(rstatix)
library(data.table)
library(ggplot2)
library(here)

# Set the root directory for all chunks
knitr::opts_knit$set(root.dir = here::here())
```

With the necessary libraries loaded, the data can be assessed. First a connection to the database is established, afterwards the tables inside can be read.

```{r load_data}
#setwd("rawData/iLand/iLand_output")
#setwd('C:/Users/ge45lep/Documents/2023_PanEuropean/r_paneurop/rawData/iLand/iLand_output')
#getwd()
# Define the path to your data using the here() function
data_path <- here("rawData/iLand/iLand_output_seeds_scenarios/20241112")
print(data_path)

# Get output dirs and run through them
lDirs <- list.dirs(path = data_path, recursive = FALSE)
#lDirs <- grep("_new", lDirs, value = TRUE)
DBTables <- list()
DBT_sap <- list()
DBT_sapd <- list()
iFiles <- 1
iDirs <- 1
bExtSeed <- FALSE

# Outer loop runs through all directories at that path and loads .sqlite files
# into a list. Also a counter (iDirs) counts the number of directories accessed
for (l_dir in lDirs) {
  lFiles <- list.files(path = l_dir, pattern = ".sqlite", full.names = TRUE)
  
  # Inner loop runs through all files (.sqlite) in a directory and loads the
  # stand table into a variable.
  # File for cluster 3_2 is skipped, due to this stand being empty in run 
  # without external seeeds
  for (lFile in lFiles) {
    l_db <- dbConnect(RSQLite::SQLite(), lFile)
    l_dfStnd <- dbReadTable(l_db, "stand")
    l_dfSapl <- dbReadTable(l_db, "sapling")
    l_dfSapD <- dbReadTable(l_db, "saplingdetail")
    dbDisconnect(l_db)
    
    bExtSeed <- !grepl("noseed", lFile)

    if (nrow(l_dfStnd) == 0) {
      next
    }
    
    l_dfStnd$filenm <- tools::file_path_sans_ext(basename(lFile))
    l_dfStnd <- l_dfStnd %>%
      dplyr::select(filenm,
                    year,
                    species,
                    volume_m3,
                    dbh_avg_cm,
                    height_avg_m,
                    basal_area_m2,
                    count_ha,
                    cohort_count_ha,
                    cohort_basal_area)
    l_dfStnd$run_nr <- iDirs
    l_dfStnd$external_seed <- bExtSeed
    
    l_dfSapl$filenm <- tools::file_path_sans_ext(basename(lFile))
    l_dfSapD$filenm <- tools::file_path_sans_ext(basename(lFile))
    # add run number
    l_dfSapl$run_nr <- iDirs
    l_dfSapD$run_nr <- iDirs
    
    DBTables[[iFiles]] <- l_dfStnd
    DBT_sap[[iFiles]] <- l_dfSapl
    DBT_sapd[[iFiles]] <- l_dfSapD
    
    iFiles <- iFiles + 1
    #dbDisconnect(l_db)
  }
  iDirs <- iDirs + 1
}
# Bring all the data together in big_data variable
OG_big_data <- bind_rows(DBTables)
OG_sapl <- bind_rows(DBT_sap)
OG_sapd <- bind_rows(DBT_sapd)

# Cleanup of some variables
rm(DBTables, l_db, l_dfStnd, iDirs, iFiles, l_dir, lDirs, lFile, lFiles, 
   bExtSeed, DBT_sapd, DBT_sap, l_dfSapl, l_dfSapD)
```

# 2. Data handling

For further processing it might be useful to split the filename into a few pieces:

-   Climate model
-   Cluster
-   With or without external seed
-   Seed input: +- % 

Organization of the iLand outputs:

-   stand by species:
-   trees \> 4 m - from output 'stand'
    -   mature trees \> 10 cm dbh
    -   juveniles \> 2m and \< 10 cm dbh
-   sapling details:
    -   \<4 m - from output 'saplings'
    -   -- 1.3, - 4 m
    -   \- has dbh -- \<1.3m - does not have dph, organized as 'cohorts' - from output 'saplings details'

in Sapling table: counts_ha and count_small_ha are calculated from the n_represented (Saplings details): n_represented = counts_ha + counts_small_ha

# Example: connect tables between iLand output formats, to get the all 3 vertical classes.

# the regeneration

| Column name       | Table          | Description                                                               |
|-------------------|----------------|---------------------------------------------------------------------------|
| count_ha          | stand          | number of trees \> 4 per ha                                               |
| basal_area_m2     | stand          | total basal area at breast height (m2)                                    |
| height_avg_m      | stand          | average tree height (m)                                                   |
| cohort_count_ha   | stand          | number of cohorts in the regeneration layer (\<4m) /ha                    |
| cohort_basal_area | stand          | basal area (m2) of saplings (\>1.3m and \<4m)                             |
| count_ha\*        | sapling        | number of represented individuals per ha (tree height \>1.3m)             |
| count_small_ha\*  | sapling        | number of represented individuals per ha (with height \<=1.3m)            |
| cohort_count_ha   | sapling        | number of cohorts in the regeneration layer (\<4m) /ha (same as in stand) |
| n_represented     | saplingDetails | number of trees that are represented by the cohort (Reineke function)     |
| dbh               | saplingDetails | diameter of the cohort (cm)                                               |
| height            | saplingDetails | height of the cohort (m)                                                  |

\*both count_ha and count_small_ha of saplings can be derived from n_represented from the sapling details

```{r data_handling}
# correct, there is 5 structural clusters in climate cluster 2, and 2 clusters in climate cluster 3



# 1. stand table
# big_data <- 
#   OG_big_data |>
#   separate_wider_delim(filenm, delim = "_", names = c("clim_model",
#                                                       "clim_scenario",
#                                                       "env_clst",
#                                                       "stnd_clst",
#                                                       "ext_seed")) #%>% 


big_data <- OG_big_data |>
 mutate(
    # Standardize: add _0 to "noseed" cases
    filenm = if_else(str_detect(filenm, "noseed$"), paste0(filenm, "_0"), filenm)
  ) |>
  separate_wider_delim(
    filenm,
    delim = "_",
    names = c("clim_model", "clim_scenario", "env_clst", "stnd_clst", "seed_flag", "seed_level")
  ) |>
  mutate(
    seed_sensitivity = case_when(
      seed_flag == "noseed" ~ 0,
      seed_level == "0" ~ 0,
      TRUE ~ as.numeric(str_remove(seed_level, "%"))
    ),
    seed_scenario = case_when(
      seed_flag == "noseed" ~ "no_seed",
      seed_level == "0" ~ "baseline",
      str_detect(seed_level, "^\\+\\d+%$") ~ "increased_seed",
      str_detect(seed_level, "^\\-\\d+%$") ~ "decreased_seed",
      TRUE ~ "unknown"
    )
  )


big_data         <- big_data %>% dplyr::select(!external_seed)
big_data$cluster <- paste(big_data$env_clst, big_data$stnd_clst, sep = "_")
big_data$clim_scenario <- as.factor(big_data$clim_scenario)
big_data$clim_model <- as.factor(big_data$clim_model)
big_data$cluster    <- as.factor(big_data$cluster)
big_data$species    <- as.factor(big_data$species)

#big_data <- big_data %>%
#  mutate(ext_seed = ifelse(external_seed, 'seed', 'noseed'))


# make new columns

# strCols <- c("volume_m3", "dbh_avg_cm", "basal_area_m2",
#              "count_ha", "cohort_count_ha", "cohort_basal_area")
# 
# w_means <- big_data %>%
#   dplyr::group_by(clim_scenario, external_seed, cluster, year, species) %>%
#   dplyr::summarise(across(strCols, list(mean = mean, sd = sd)),.groups = "drop")

# 2. sapling table
# big_sapl <- OG_sapl |>
#   separate_wider_delim(filenm, delim = "_", names = c("clim_model",
#                                                       "clim_scenario",
#                                                       "env_clst",
#                                                       "stnd_clst",
#                                                       "ext_seed"))
# big_sapl$cluster <- paste(big_sapl$env_clst, big_sapl$stnd_clst, sep = "_")
# big_sapl$clim_model <- as.factor(big_sapl$clim_model)
# big_sapl$clim_scenario <- as.factor(big_sapl$clim_scenario)
# big_sapl$cluster <- as.factor(big_sapl$cluster)
# big_sapl$species <- as.factor(big_sapl$species)
# big_sapl$ext_seed <- as.factor(big_sapl$ext_seed)
# big_sapl <- big_sapl %>% dplyr::select(!c(env_clst, stnd_clst, ru, rid))

# strCols <- c("count_ha", "count_small_ha", "cohort_count_ha", "height_avg_m",
#              "age_avg", "LAI")
# 
# w_means_sapl <- big_sapl %>%
#   dplyr::group_by(clim_scenario, ext_seed, cluster, year, species) %>%
#   dplyr::summarise(across(strCols, list(mean = mean, sd = sd)),.groups = "drop")

# 3. sapling details table
# big_sapD <- OG_sapd |>
#   separate_wider_delim(filenm, delim = "_", names = c("clim_model",
#                                                       "clim_scenario",
#                                                       "env_clst",
#                                                       "stnd_clst",
#                                                       "ext_seed"))

big_sapD <- 
  OG_sapd |>
 mutate(
    # Standardize: add _0 to "noseed" cases
    filenm = if_else(str_detect(filenm, "noseed$"), paste0(filenm, "_0"), filenm)
  ) |>
  separate_wider_delim(
    filenm,
    delim = "_",
    names = c("clim_model", "clim_scenario", "env_clst", "stnd_clst", "seed_flag", "seed_level")
  ) |>
  mutate(
    seed_sensitivity = case_when(
      seed_flag == "noseed" ~ 0,
      seed_level == "0" ~ 0,
      TRUE ~ as.numeric(str_remove(seed_level, "%"))
    ),
    seed_scenario = case_when(
      seed_flag == "noseed" ~ "no_seed",
      seed_level == "0" ~ "baseline",
      str_detect(seed_level, "^\\+\\d+%$") ~ "increased_seed",
      str_detect(seed_level, "^\\-\\d+%$") ~ "decreased_seed",
      TRUE ~ "unknown"
    )
  )



big_sapD$cluster        <- paste(big_sapD$env_clst, big_sapD$stnd_clst, sep = "_")
big_sapD$clim_model     <- as.factor(big_sapD$clim_model)
big_sapD$clim_scenario  <- as.factor(big_sapD$clim_scenario)
big_sapD$cluster        <- as.factor(big_sapD$cluster)
big_sapD$species        <- as.factor(big_sapD$species)
#big_sapD$ext_seed       <- as.factor(big_sapD$ext_seed)
big_sapD <- big_sapD %>% dplyr::select(!c(env_clst, stnd_clst, ru, rid))

# strCols <- c("age", "height", "dbh", "n_represented")
# 
# w_means_sapD <- big_sapD %>%
#   dplyr::group_by(clim_scenario, ext_seed, cluster, year, species) %>%
#   dplyr::summarise(across(strCols, list(mean = mean, sd = sd)),.groups = "drop")



# put all data together
# check if cohorst make sense!
# dplyr::filter(year == 0 & cluster == "1_1" & clim_model == "NCC" & clim_scenario == "HISTO")
#head(big_data)
#head(big_sapl)
#head(big_sapD)


```

Prepare new vertical classification and merge the data:

-   stand: for trees \> 4 m tall
-   sapling: for trees \< 4 m tall

```{r final_tab_vert_cat}

# split in vertical classes: 
#  - big data: 
#    - mature
#    - juveniles
#  - saplD:
#    - juveniles
#    - saplings
#    - seedlings (< 20 cm, remove)


# Start with the stand data: 
# - ignore cohors (those are small regeneration)
# - split into mature and juveniles, sums up by species

# From sapl Details: 
# - split the cohorts: split into vertical classes (seedlings (<0.2m), saplings (0.2-2m), juveniles 2m-4m)
# sum up by vertical classes, get count_ha and dbh for basal area
# no need to do this with the saplings table, as i have the counts and BA from teh saplings details (saplD)

# classification seems correct
# out_big_test_juv %>% 
#   dplyr::filter(category == 'mature') %>% 
#   summary()


# Trees  taller > 4 m
out_big <- 
  big_data %>% 
  # get vert_classes
   mutate(category = case_when(
     height_avg_m > 2 & dbh_avg_cm < 10 ~ "juveniles",
     dbh_avg_cm >= 10 ~ "mature",
    
   # dbh_avg_cm < 10 ~ "juveniles",
   
    TRUE ~ "unclassified"  # thse are very young regeneration, that is less then 20 cm tall, so I can remove
  )) %>% 
  #  rename(ext_seed = external_seed) %>% 
  #  head()
 # dplyr::filter(category != 'seedlings') %>% 
 # mutate(r=dbh/100/2, # convert cm to m
#         BA = pi*r^2,
#         basal_area_m2 = BA*n_represented) %>% # how to calculate cohort basal area???
  group_by(year, species, clim_model, clim_scenario, run_nr, cluster, category, seed_flag, seed_level, seed_sensitivity , seed_scenario) %>% #external_seed
  summarise(count_ha = sum(count_ha ),  # summarize by run, and by species and vertical class
            basal_area_m2 = sum(basal_area_m2)) %>%
  ungroup() %>% 
  dplyr::filter(category %in% c('mature', 'juveniles'))

# Trees < 4m of height
out_sapD <- big_sapD %>% 
  # get vert_classes
   mutate(category = case_when(
    height > 2 ~ "juveniles",
    height >= 0.2 & height <= 2 ~ "saplings",
     height < 0.2 ~ "seedlings",
   # height_avg_m > 2 & dbh_avg_cm <= 10 ~ "juveniles",
    TRUE ~ "unclassified"  # those are teh very young regeneration, that is less then 20 cm tall, so I can remove
  )) %>% 
  dplyr::filter(category != 'seedlings') %>%  # remove seedling less then 20 cm
  mutate(r=dbh/100/2, # convert cm to m
         BA = pi*r^2,
         basal_area_m2 = BA*n_represented) %>% # how to calculate cohort basal area???
  group_by(year, species, clim_model, clim_scenario, run_nr, cluster, category, seed_flag, seed_level, seed_sensitivity , seed_scenario) %>% 
  summarise(count_ha = sum(n_represented),  # summarize by run, and by species and vertical class
            basal_area_m2 = sum(basal_area_m2)) %>%
  ungroup()
  



out_sapD %>% 
  dplyr::filter(cluster == '1_1')
head(out_sapD)
head(out_big)


```

Test subsets: do I get the data right???

```{r, test-subsets}

cluster_name = '1_1'

# subset on one landscape, to make sure I got the cohorts right!
big_data_sub <- out_big %>%
  dplyr::filter(year == 0 & cluster == cluster_name & clim_model == "NCC" & clim_scenario == "HISTO" & run_nr == 1 ) # & ext_seed == 'noseeed'

# thesea are mature trees
View(big_data_sub)
# big_sapl_sub <- big_sapl %>% 
#   dplyr::filter(year == 0 & cluster == "1_1" & clim_model == "NCC" & clim_scenario == "HISTO" & ext_seed == 'noseed' & run_nr == 1 )
# View(big_sapl_sub)

big_sapD_sub <- 
  out_sapD %>% 
  dplyr::filter(year == 0 & cluster == cluster_name & clim_model == "NCC"  & clim_scenario == "HISTO"& ext_seed == 'noseed' & run_nr == 1 ) #   & clim_scenario == "HIST0"
View(big_sapD_sub)
#big_data_sub <- 
# clean environment
#rm(strCols)
```

```{r export-table}
df_simulated_merged <- rbind(out_sapD, out_big)

# aggregate values based on vert category: as I have juveniles from stand and saplings iland tables
df_simulated <- df_simulated_merged %>% 
  group_by(year, species, clim_model, clim_scenario, run_nr, cluster, category, seed_flag, seed_level, seed_sensitivity , seed_scenario) %>% 
  summarise(count_ha = sum(count_ha),
            basal_area_m2 = sum(basal_area_m2))


# juveniles can be from saplings and stand level class: merge them together
head(df_simulated)

df_simulated %>% 
  dplyr::filter(cluster == cluster_name & year == 1) %>% 
  View()
  #summary() %>% 
  




getwd()

fwrite(df_simulated, 'outTable/df_simulated.csv')

```



Description of columns:

| Column name            | Description                                                                    |
|------------------------|--------------------------------------------------------------------------------|
| external_seed          | Wheather external seed was considered in the simulation                        |
| cluster                | To which of the 12 clusters this line belongs                                  |
| year                   | simulated year (0 is initial state, 30 final state)                            |
| species                | tree species the data relates to                                               |
| volume_m3_mean         | volume of species biomass (mean)                                               |
| volume_m3_sd           | volume of species biomass (standard deviation)                                 |
| dbh_avg_cm_mean        | average dbh of trees (\>4m height) (mean)                                      |
| dbh_avg_cm_sd          | average dbh of trees (\>4m height) (standard deviation)                        |
| basal_area_m2_mean     | total basal area at breast height (\>4m height) (mean)                         |
| basal_area_m2_sd       | total basal area at breast height (\>4m height) (standard deviation)           |
| count_ha_4_mean        | tree count (living, \>4m height) per ha (mean)                                 |
| count_ha_4_sd          | tree count (living, \>4m height) per ha (standard deviation)                   |
| co_cnt_ha_mean         | number of cohorts in the regeneration layer (\<4m) per ha (mean)               |
| co_cnt_ha_sd           | number of cohorts in the regeneration layer (\<4m) per ha (standard deviation) |
| cohort_basal_area_mean | basal area of saplings (\>1.3m and \<4m) (mean)                                |
| cohort_basal_area_sd   | basal area of saplings (\>1.3m and \<4m) (standard deviation)                  |
| cnt_ha_13_mean         | number of represented individuals per ha (\>1.3m height) (mean)                |
| cnt_ha_13_sd           | number of represented individuals per ha (\>1.3m height) (standard deviation)  |
| count_small_ha_mean    | number of represented individuals per ha (\<=1.3m height) (mean)               |
| count_small_ha_sd      | number of represented individuals per ha (\<=1.3m height) (standard deviation) |
| cohort_count_ha_mean.y | number of cohorts per ha (same as co_cnt_ha_mean)                              |
| cohort_count_ha_sd.y   | number of cohorts per ha (same as co_cnt_ha_sd)                                |
| height_avg_m_mean      | arithmetic average height of the cohorts (m) (mean)                            |
| height_avg_m_sd        | arithmetic average height of the cohorts (m) (sd)                              |
| age_avg_mean           | arithmetric average age of the sapling cohorts (years) (mean)                  |
| age_avg_sd             | arithmetric average age of the sapling cohorts (years) (standard deviaion)     |
| age_avg_sd             | arithmetric average age of the sapling cohorts (years) (standard deviaion)     |
| n_represented_mean     | number of trees that are represented by the cohort (Reineke function) (mean)   |
| n_represented_sd       | number of trees that are represented by the cohort (Reineke function) (sd)     |
| dbh_mean               | diameter of the cohort in cm (mean)                                            |
| dbh_sd                 | diameter of the cohort in cm (sd)                                              |
| height_mean            | height of the cohort in m (mean)                                               |
| height_sd              | height of the cohort in m (sd)                                                 |
| age_mean               | age of the cohort (mean)                                                       |
| age_sd                 | age of the cohort (sd)                                                         |

: column descriptions
